legacy/ALoRA-MoE/requirements.txt
torch>=2.2
torchvision>=0.17
einops>=0.8
numpy>=1.24
scipy>=1.10
nibabel>=5.2
monai>=1.3
scikit-image>=0.22
tqdm>=4.66
pyyaml>=6.0
